u32 -> i64

+++ src/domain/video/repository.rs
@@ -7,9 +7,9 @@ pub trait VideoRepository {
     async fn list_by_channel(
         &self,
         channel_id: String,
-        limit: u32,
-        offset: u32,
+        limit: i64,
+        offset: i64,
     ) -> Result<Vec<Video>>;
-    async fn list(&self, limit: u32, offset: u32) -> Result<Vec<Video>>;
+    async fn list(&self, limit: i64, offset: i64) -> Result<Vec<Video>>;
     async fn create(&self, video: DraftVideo) -> Result<()>;
 }
